Version 675
An user found a problem with the degradation output in the 'output.sed' file (consistently output as zero).
The model was computing this correctly, but the variable printing to the output file was incorrect.
Changes to the rtsed routine were made for this fix.  (see Jeewa 08/28/2019 email)

main.f - revision and date change
rtsed.f -
old:         rchdy(56,jrch) = deg2
new:       rchdy(56,jrch) = deg1 + deg2